ft: fix warning

+++ src/cairo-ft-font.c
@@ -2493,7 +2493,7 @@ _cairo_index_to_glyph_name (void	         *abstract_font,
      * we fall back to searching the entire array.
      */
 
-    if (glyph_index < num_glyph_names &&
+    if ((long)glyph_index < num_glyph_names &&
 	strcmp (glyph_names[glyph_index], buffer) == 0)
     {
 	*glyph_array_index = glyph_index;
